Push up initial things to make example narrative build
I've grabbed the data available from the public build up as of today (Jan 8, 2020), but taken the Snakefile and config directories from the private repo. My goal here is to make a build that we can use for making narratives for the manuscript. This means that it needs the clade labeling steps from the private build, but needs to not have epi IDs.