Changed file name to try to fix loading issue on website
Currently, if you go directly to https://nextstrain.org/community/blab/ebola-narrative-ms, the dataset loads no problem. However, if you're in directory area where you can access other datasets as well as the narratives, this loading isn't working because the URL routes to: https://nextstrain.org/community/blab/ebola-narrative-ms/ebola-narrative-ms. I believe that this might be because the JSON file name for the dataset exactly matches the repo name, so I'm trying out appending a bit more to the dataset name to get around this.